Renamed some files and organized them in folders
