+++ lib/authentication/services/authentication_service.dart
@@ -4,9 +4,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:smarttodo/authentication69/database69.dart';
-import 'package:smarttodo/authentication69/wrapper69.dart';
-import 'package:smarttodo/models69/app_user.dart';
+import 'package:smarttodo/authentication/services/database_service.dart';
+import 'package:smarttodo/authentication/services/wrapper.dart';
+import 'package:smarttodo/models/app_user.dart';
 
 class AuthenticationService {
   final FirebaseAuth _firebaseAuth;
