+++ lib/homepage.dart
@@ -7,7 +7,7 @@ import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
-import 'package:smarttodo/authentication69/authentication_service69.dart';
+import 'package:smarttodo/authentication/services/authentication_service.dart';
 import 'package:smarttodo/constants.dart';
 import 'package:url_launcher/url_launcher.dart';
