+++ lib/models/app_user.dart
